+++ backend/internal/store/store.go
@@ -474,7 +474,7 @@ func (s *Store) MuteByMerchant(ctx context.Context, pattern, reason string) erro
 // CategorizeMerchant atomically updates all transactions with the given merchant_info
 // (exact case-sensitive equality match, not substring) and upserts a user_locked entry
 // in merchant_categories for future scans. Returns the number of transaction rows updated.
-func (s *Store) CategorizeMerchant(ctx context.Context, merchant, category, bucket string) (int, error) {
+func (s *Store) CategorizeMerchant(ctx context.Context, merchant, category, bucket string) (int64, error) {
 	return s.community.CategorizeMerchant(ctx, merchant, category, bucket)
 }
 
@@ -524,7 +524,7 @@ func (s *Store) SeedMCCCodes(ctx context.Context, entries []MCCEntry) error {
 
 // SeedMerchantCategories upserts community merchant fragment mappings, skipping
 // rows where user_locked = true (user has explicitly modified the entry).
-func (s *Store) SeedMerchantCategories(ctx context.Context, entries []MerchantCategoryEntry) (int, error) {
+func (s *Store) SeedMerchantCategories(ctx context.Context, entries []MerchantCategoryEntry) (int64, error) {
 	return s.community.SeedMerchantCategories(ctx, entries)
 }
